requirements.txt
imutils
mss
numpy
opencv_python
Pillow
pytesseract
